requirements.txt
google-genai>=0.3.0
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
